Problem on C

+++ College/Input-Output/P3.c
@@ -3,11 +3,11 @@
 void main(){
 
 //sum and avg of 3 nos
-    int i,j,k,sum,avg;
+    float i,j,k,sum,avg;
     printf("Enter value of 3 nos.");
-    scanf("%d%d%d",&i,&j,&k);
+    scanf("%f%f%f",&i,&j,&k);
     sum=i+j+k;
     avg=sum/3;
-    printf("Average of 3 nos: %d", avg );
+    printf("Average of 3 nos: %f", avg );
 
 }
